Add proxy to details, add error handling
I added the proxy into the `product.details`, which was not present. I also added error throwing for `products.search` if the query is undefined. I removed the unnecessary try catch statements and implemented the error handling from utils, and allowed the user to input a previously fetched product in `product.details`.